refactor(datasets): explicitly call remove in sub stores
Avoids Pinia $onAction magic on delete calls.  components
call datasetStore.remove, rather than imageStore or dicomStore delete.
